requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3
pillow==10.0.1
python-socketio==5.9.0
flask-socketio==5.3.6
